Working on a reimplementation of the bookmarks manager...
This isn't finished yet, but the basic idea is to abandon UIDocument in favor
of writing the bookmarks file directly (coordinated of course). This has the
benefit of faster sync (because we don't have to wait for the document to auto-
save) and (hopefully) a cleaner method of dealing with the bookmarks file 
disappearing from iCloud (when the user deleted it from the system preferences).